move authService to src/services.
